+++ CHANGELOG.md
@@ -2,7 +2,13 @@
 
 _The whole finder CHANGELOG_
 
-## Newest Version : prob 0.3.16
+## Newest prob 0.3.17
+
+- added regex support
+- added mcapp minecraft world dashboard
+- fixed buildcheck workflow
+
+## 0.3.16 - 08.09.2026
 
 - added cache creation
 - added config editor